requirements.txt
discord.py==1.2.3
mojang-api==2.1.0
requests==2.22.0